Fix include paths after merging master (EncodedIriManager moved in #3146)
